Moved price and priceExcludingTax from Store reducer to Delivery reducer
